# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Testigos API",
        "description": "API para plataforma de crowdfunding legal - Conecta víctimas, abogados e inversores para financiar casos legales",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/cases/{id}": {
            "get": {
                "tags": [
                    "Cases"
                ],
                "summary": "Obtener detalles de un caso",
                "operationId": "c6b78d4ceacff3b256c07e796f847515",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del caso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del caso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Case"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Caso no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cases"
                ],
                "summary": "Actualizar caso",
                "operationId": "68c8837fdd8eaad630769d3d535dcdde",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del caso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCaseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Caso actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Case"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Caso no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cases"
                ],
                "summary": "Eliminar caso",
                "operationId": "2c10dd2f14d5cfba4a4c609aa83d355e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del caso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Caso eliminado exitosamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Caso no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/transactions/case/{caseId}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Obtener todas las transacciones de un caso específico",
                "operationId": "71c774375e25ca4688ca0552951b232f",
                "parameters": [
                    {
                        "name": "caseId",
                        "in": "path",
                        "description": "ID del caso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transacciones del caso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "transactions": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Transaction"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado para ver este caso"
                    },
                    "404": {
                        "description": "Caso no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Case": {
                "required": [
                    "id",
                    "title",
                    "description",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Despido injustificado durante licencia médica"
                    },
                    "description": {
                        "type": "string",
                        "example": "Fui despedida de mi trabajo mientras estaba con licencia médica..."
                    },
                    "victim_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "lawyer_id": {
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "pending",
                            "published",
                            "funding",
                            "in_progress",
                            "completed",
                            "closed"
                        ],
                        "example": "published"
                    },
                    "category": {
                        "type": "string",
                        "example": "despido_injustificado"
                    },
                    "company": {
                        "type": "string",
                        "example": "Retail Chile S.A."
                    },
                    "funding_goal": {
                        "type": "number",
                        "format": "decimal",
                        "example": 1500000
                    },
                    "current_funding": {
                        "type": "number",
                        "format": "decimal",
                        "example": 750000
                    },
                    "success_rate": {
                        "type": "number",
                        "format": "decimal",
                        "example": 85
                    },
                    "expected_return": {
                        "type": "number",
                        "format": "decimal",
                        "example": 30
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-01"
                    },
                    "legal_analysis": {
                        "type": "string"
                    },
                    "evaluation_data": {
                        "type": "object"
                    },
                    "lawyer_evaluation_fee": {
                        "type": "number",
                        "format": "decimal",
                        "example": 500000
                    },
                    "lawyer_success_fee_percentage": {
                        "type": "number",
                        "format": "decimal",
                        "example": 15
                    },
                    "lawyer_fixed_fee": {
                        "type": "number",
                        "format": "decimal",
                        "example": 2000000
                    },
                    "lawyer_total_compensation": {
                        "type": "number",
                        "format": "decimal",
                        "example": 2500000
                    },
                    "lawyer_paid_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "outcome": {
                        "type": "string",
                        "enum": [
                            "won",
                            "lost",
                            "settled",
                            "dismissed"
                        ],
                        "example": "won",
                        "nullable": true
                    },
                    "amount_recovered": {
                        "type": "number",
                        "format": "decimal",
                        "example": 15000000,
                        "nullable": true
                    },
                    "legal_costs": {
                        "type": "number",
                        "format": "decimal",
                        "example": 2000000
                    },
                    "outcome_description": {
                        "type": "string",
                        "nullable": true
                    },
                    "resolution_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-11-15",
                        "nullable": true
                    },
                    "closed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CreateCaseRequest": {
                "required": [
                    "title",
                    "description",
                    "category"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Despido injustificado"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descripción detallada del caso..."
                    },
                    "category": {
                        "type": "string",
                        "example": "despido_injustificado"
                    },
                    "company": {
                        "type": "string",
                        "example": "Empresa ABC S.A."
                    },
                    "funding_goal": {
                        "type": "number",
                        "example": 1500000
                    }
                },
                "type": "object"
            },
...
            "Transaction": {
                "required": [
                    "id",
                    "transaction_id",
                    "type",
                    "amount",
                    "direction",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "transaction_id": {
                        "type": "string",
                        "example": "TXN-ABC123DEF456"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "investment",
                            "platform_commission",
                            "success_commission",
                            "lawyer_payment",
                            "investor_return",
                            "withdrawal",
                            "gateway_fee",
                            "refund"
                        ],
                        "example": "investment"
                    },
                    "case_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "investment_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "amount": {
                        "type": "number",
                        "format": "decimal",
                        "example": 1000000
                    },
                    "currency": {
                        "type": "string",
                        "example": "CLP"
                    },
                    "direction": {
                        "type": "string",
                        "enum": [
                            "in",
                            "out"
                        ],
                        "example": "in"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "processing",
                            "completed",
                            "failed",
                            "cancelled"
                        ],
                        "example": "completed"
                    },
                    "payment_gateway": {
                        "type": "string",
                        "example": "transbank",
                        "nullable": true
                    },
                    "gateway_transaction_id": {
                        "type": "string",
                        "example": "TBK-789456123",
                        "nullable": true
                    },
                    "gateway_fee": {
                        "type": "number",
                        "format": "decimal",
                        "example": 25000
                    },
                    "gateway_response": {
                        "type": "object",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Inversión en caso #1"
                    },
                    "metadata": {
                        "type": "object",
                        "nullable": true
                    },
                    "processed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "completed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}